logout + 한글깨짐 수정 (#71)
