+++ app/src/main/java/com/kosa/thirdprojectfront/RetrofitBuilder.kt
@@ -10,7 +10,7 @@ object RetrofitBuilder {
 
     init {
         val retrofit = Retrofit.Builder()
-            .baseUrl("http://192.168.0.125:8080/")
+            .baseUrl("http://192.168.0.101:8080/")
             .addConverterFactory(GsonConverterFactory.create(gson))
             .build()
         api = retrofit.create(API::class.java)
