+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -4,9 +4,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/loginFragment"
-        android:icon="@drawable/ic_login_24"
-        android:title="LOGIN"
+        android:id="@+id/logoutFragment"
+        android:icon="@drawable/ic_baseline_logout_24"
+        android:title="LOGOUT"
         app:showAsAction="always"/>
